kvcoord: extract txn correctness tests to subpkg
TestTxnDBDirtyWriteAnomaly takes minutes to run, so it's likely to cause test
timeouts when grouped into an unfortunate shard with other long-running tests
in this package.

PS: I'm not sure why bazel didn't pick up on the `kvcoord` import; I added
it manually (via #keep).

Release note: none
Epic: none

x
