+++ pkg/BUILD.bazel
@@ -207,6 +207,7 @@ ALL_TESTS = [
     "//pkg/keyvisualizer/spanstatsconsumer:spanstatsconsumer_test",
     "//pkg/kv/bulk/bulkpb:bulkpb_test",
     "//pkg/kv/bulk:bulk_test",
+    "//pkg/kv/kvclient/kvcoord/txncorrectnesstest:txncorrectnesstest_test",
     "//pkg/kv/kvclient/kvcoord:kvcoord_disallowed_imports_test",
     "//pkg/kv/kvclient/kvcoord:kvcoord_test",
     "//pkg/kv/kvclient/kvstreamer:kvstreamer_test",
@@ -1381,6 +1382,7 @@ GO_TARGETS = [
     "//pkg/kv/bulk:bulk",
     "//pkg/kv/bulk:bulk_test",
     "//pkg/kv/kvbase:kvbase",
+    "//pkg/kv/kvclient/kvcoord/txncorrectnesstest:txncorrectnesstest_test",
     "//pkg/kv/kvclient/kvcoord:kvcoord",
     "//pkg/kv/kvclient/kvcoord:kvcoord_test",
     "//pkg/kv/kvclient/kvstreamer:kvstreamer",
