+++ pkg/kv/kvclient/kvcoord/BUILD.bazel
@@ -145,7 +145,6 @@ go_test(
         "txn_coord_sender_savepoints_test.go",
         "txn_coord_sender_server_test.go",
         "txn_coord_sender_test.go",
-        "txn_correctness_test.go",
         "txn_interceptor_committer_test.go",
         "txn_interceptor_heartbeater_test.go",
         "txn_interceptor_metric_recorder_test.go",
@@ -182,7 +181,6 @@ go_test(
         "//pkg/kv/kvpb/kvpbmock",
         "//pkg/kv/kvserver",
         "//pkg/kv/kvserver/closedts",
-        "//pkg/kv/kvserver/concurrency",
         "//pkg/kv/kvserver/concurrency/isolation",
         "//pkg/kv/kvserver/concurrency/lock",
         "//pkg/kv/kvserver/kvserverbase",
