+++ pkg/kv/kvclient/kvcoord/txncorrectnesstest/BUILD.bazel
@@ -1,0 +1,31 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_test")
+
+go_test(
+    name = "txncorrectnesstest_test",
+    srcs = [
+        "main_test.go",
+        "txn_correctness_test.go",
+    ],
+    deps = [
+        "//build/bazelutil:noop",
+        "//pkg/base",
+        "//pkg/kv",
+        "//pkg/kv/kvclient/kvcoord",  # keep
+        "//pkg/kv/kvserver",
+        "//pkg/kv/kvserver/concurrency",
+        "//pkg/kv/kvserver/concurrency/isolation",
+        "//pkg/kv/kvserver/kvserverbase",
+        "//pkg/security/securityassets",
+        "//pkg/security/securitytest",
+        "//pkg/server",
+        "//pkg/storage/enginepb",
+        "//pkg/testutils/localtestcluster",
+        "//pkg/testutils/serverutils",
+        "//pkg/util/leaktest",
+        "//pkg/util/log",
+        "//pkg/util/randutil",
+        "//pkg/util/syncutil",
+        "@com_github_cockroachdb_errors//:errors",
+        "@com_github_stretchr_testify//require",
+    ],
+)
